+++ src/utils.cc
@@ -24,6 +24,8 @@
    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
+#include <stdint.h> // int64_t
+
 #include <sched.h>
 #include <unistd.h>
 #include <errno.h>
@@ -38,10 +40,10 @@
 namespace hogl
 {
 
-int setaffinity(pthread_t thread_id, int core_id) {
+int setaffinity(pthread_t thread_id, int64_t core_id) {
    // Do nothing in case of core_id -1 or < 0
    if (core_id < 0) { return 0; }
-   int num_cores = sysconf(_SC_NPROCESSORS_ONLN);
+   int64_t num_cores = sysconf(_SC_NPROCESSORS_ONLN);
    if (core_id < 0 || core_id >= num_cores)
       return EINVAL;
